+++ src/main/java/top/howiehz/halo/plugin/extra/api/service/PostStatsDataCacheManager.java
@@ -6,6 +6,7 @@ import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Component;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicReference;
+import java.math.BigInteger;
 
 /**
  * Stats Data Manager - 统计数据管理器。
@@ -17,13 +18,12 @@ import java.util.concurrent.atomic.AtomicReference;
 public class PostStatsDataCacheManager {
 
     // 文章发布版本字数统计缓存 / Total & per-post caches for release version
-    private final AtomicReference<Integer> totalReleasePostWordCount = new AtomicReference<>(null);
-    final ConcurrentHashMap<String, Integer> releasePostWordCounts =
-        new ConcurrentHashMap<>();
+    private final AtomicReference<BigInteger> totalReleasePostWordCount =
+        new AtomicReference<>(null);
+    final ConcurrentHashMap<String, BigInteger> releasePostWordCounts = new ConcurrentHashMap<>();
     // 文章草稿版本字数统计缓存 / Total & per-post caches for draft version
-    private final AtomicReference<Integer> totalDraftPostWordCount = new AtomicReference<>(null);
-    final ConcurrentHashMap<String, Integer> draftPostWordCounts =
-        new ConcurrentHashMap<>();
+    private final AtomicReference<BigInteger> totalDraftPostWordCount = new AtomicReference<>(null);
+    final ConcurrentHashMap<String, BigInteger> draftPostWordCounts = new ConcurrentHashMap<>();
 
     /**
      * 获取缓存的总字数。
@@ -33,7 +33,7 @@ public class PostStatsDataCacheManager {
      * @return 总字数，未缓存时返回 null / Total word count, returns null if not cached
      */
     @Nullable
-    public Integer getCachedTotalWordCount(boolean isDraft) {
+    public BigInteger getCachedTotalWordCount(boolean isDraft) {
         return isDraft ? totalDraftPostWordCount.get() : totalReleasePostWordCount.get();
     }
 
@@ -46,7 +46,7 @@ public class PostStatsDataCacheManager {
      * @return 文章字数，未缓存时返回 null / Post word count, returns null if not cached
      */
     @Nullable
-    public Integer getCachedPostWordCount(String postName, boolean isDraft) {
+    public BigInteger getCachedPostWordCount(String postName, boolean isDraft) {
         return isDraft ? draftPostWordCounts.get(postName) : releasePostWordCounts.get(postName);
     }
 
@@ -57,7 +57,7 @@ public class PostStatsDataCacheManager {
      * @param count 总字数 / Total word count
      * @param isDraft 是否为草稿版本 / Whether it's draft version
      */
-    public void setTotalPostWordCount(int count, boolean isDraft) {
+    public void setTotalPostWordCount(BigInteger count, boolean isDraft) {
         if (isDraft) {
             totalDraftPostWordCount.set(count);
             log.debug("Set total DRAFT word count cache: {}", count);
@@ -75,7 +75,7 @@ public class PostStatsDataCacheManager {
      * @param count 字数 / Word count
      * @param isDraft 是否为草稿版本 / Whether it's draft version
      */
-    public void setPostWordCount(String postName, int count, boolean isDraft) {
+    public void setPostWordCount(String postName, BigInteger count, boolean isDraft) {
         if (isDraft) {
             draftPostWordCounts.put(postName, count);
             log.debug("Set DRAFT word count cache for post {}: {}", postName, count);
